+++ src/default-rules/circular-dependencies/types.ts
@@ -1,0 +1,1 @@
+export type Mapping = (source: string, target: string) => string;